Util functions separated into new module.
