+++ emailNewPosts.js
@@ -8,6 +8,10 @@ const wrap80 = require('wordwrap')(80);
 const yaml = require('js-yaml');
 
 const smtpTransport = require('./smtpTransport');
+const utils = require('./utils');
+const readFileAsString = utils.readFileAsString;
+const readFileAsArray = utils.readFileAsArray;
+const writeArrayToFile = utils.writeArrayToFile;
 
 const configFile = fs.readFileSync('./_config.yml', 'utf8');
 const config = yaml.safeLoad(configFile).server;
@@ -15,28 +19,6 @@ const config = yaml.safeLoad(configFile).server;
 const postsDirectory = path.join(process.cwd(), 'source/_posts');
 const emailLogFilename = path.join(process.cwd(), 'email_update_log');
 const mailingListFilename = path.join(process.cwd(), 'mailing_list');
-
-function readFileAsString (filename, callback) {
-  fs.readFile(filename, function (err, buffer) {
-    if (err) {
-      return callback(err);
-    }
-    callback(null, buffer.toString());
-  });
-}
-
-function readFileAsArray (filename, callback) {
-  readFileAsString(filename, function (err, contents) {
-    if (err) {
-      return callback(err);
-    }
-    callback(null, contents.split('\n'));
-  });
-}
-
-function writeArrayToFile (filename, array, callback) {
-  fs.writeFile(filename, array.join('\n'), callback);
-}
 
 function sendEmail (emailData, address, callback) {
   let mailOptions = {
